# hugolib: Simplify replaceShortcodeTokens