Simplify the code of `replaceShortcodeTokens`:
* Delete some unclear variables like `j` and `k`. It was hard to understand the code
* Use `bytes.Buffer` from pool for `source`, so this function doesn't allocate a lot of memory
* Delete `prefix` argument, why in real cases it's always `shortcodePlaceholderPrefix`

Earlier replaceShortCodeTokens consumed ~11% of all memory hugo consumes for `hugo -s docs`:
![2017-05-07 21 26 32](https://cloud.githubusercontent.com/assets/13235519/25785120/15e4c712-3379-11e7-81eb-bb0b6d738238.png)
Now it's only ~1%

Benchmarks:
Best result before:
```
Average time per operation: 428ms
Average memory allocated per operation: 127350kB
Average allocations per operation: 2137892
```

After:
```
Average time per operation: 425ms
Average memory allocated per operation: 113991kB
Average allocations per operation: 2137522
```
---
**UPDATE: Tests are passed**
The only problem now is, what it can't pass 2 tests yet:
```
$ go test ./hugolib
Source changed content/sect/doc2.en.md
Source changed content/new1.en.md
Source changed content/new2.en.md
Source changed content/new1.fr.md
Source changed content/sect/doc1.en.md
Source changed content/new1renamed.en.md
Source changed content/new1.en.md
Template changed layouts/_default/single.html
i18n changed i18n/fr.yaml
Template changed layouts/shortcodes/shortcode.html
ERROR 2017/05/07 23:06:35 missing : template: missing:1: function "funcdoesnotexists" not defined
ERROR 2017/05/07 23:06:35 init : template: missing:1: function "funcdoesnotexists" not defined
ERROR 2017/05/07 23:06:35 asciidoctor / asciidoc not found in $PATH: Please install.
                  Leaving AsciiDoc content unrendered.
ERROR 2017/05/07 23:06:35 rst2html / rst2html.py not found in $PATH: Please install.
                  Leaving reStructuredText content unrendered.
--- FAIL: TestShortcodesInSite (0.04s)
        Error Trace:    testhelpers_test.go:42
	            	shortcode_test.go:561
	Error:      	Should be true
	Messages:   	File no match for
	            	"<p>a</p>\n\n<p>b<br />\nc\nd</p>\n\n<p>e</p>\n" in
	            	"public/sect/doc2/index.html":
	            	<p>a</p>

	            	<p>b
	            	c
	            	d</p>

	            	<p>e</p>
--- FAIL: TestInnerSCWithAndWithoutMarkdown (0.02s)
	shortcode_test.go:90: Shortcode render didn't match. got
		"<div><h1 id=\"more-here\">More Here</h1>\n\n<p><a href=\"http://spf13.com\">link</a> and text</p>\n</div>\n\n<p>And then:</p>\n\n<div>\n# More Here\n\nThis is **plain** text.\n\n</div>" but expected
		"<div><h1 id=\"more-here\">More Here</h1>\n\n<p><a href=\"http://spf13.com\">link</a> and text</p>\n</div>\n\n<p>And then:</p>\n\n<p><div>\n# More Here\n\nThis is **plain** text.\n\n</div>"
ERROR 2017/05/07 23:06:35 No shortcode param at .Get 1 in page simple.md, have params: [47238zzb]
WARNING: Site config's rssURI is deprecated and will be removed in a future release. Set baseName in outputFormats.RSS.
ERROR 2017/05/07 23:06:36 Two or more menu items have the same name/identifier in Menu "m1": "n2".
Rename or set an unique identifier.
ERROR 2017/05/07 23:06:36 Two or more menu items have the same name/identifier in Menu "m1": "i2".
Rename or set an unique identifier.
ERROR 2017/05/07 23:06:36 Two or more menu items have the same name/identifier in Menu "m1": "i2".
Rename or set an unique identifier.
ERROR 2017/05/07 23:06:36 Two or more menu items have the same name/identifier in Menu "m1": "n2".
Rename or set an unique identifier.
Skipping shortname timezone tests.
ERROR 2017/05/07 23:06:36 Failed to parse date '2010-05-02_15:29:31+08:00' in page page/with/invalid/date
WARNING: Page's Now is deprecated and will be removed in a future release. Use now (the template func).
ERROR 2017/05/07 23:06:36 page content/post/broken.md has both draft and published settings in its frontmatter. Using draft.
ERROR 2017/05/07 23:06:36 both draft and published parameters were found in page's frontmatter
ERROR 2017/05/07 23:06:36 unable to read frontmatter at filepos 45: EOF
FAIL
FAIL	github.com/spf13/hugo/hugolib	1.882s
```
First tests added by https://github.com/spf13/hugo/commit/1ce184b7f1df10ad8815dfe891e053eb04f099ea and it says, what they should not be passed. And they are not passed in my PR :(
In second test `<p>` lacks. I don't know why.

TODO:
- [x] Pass tests
- [ ] Add more comments

/cc @bep