@bep 2017-05-07
I'm not sure how you arrived at your conclusion, but the benchmark for that function shows a pretty clear winner in the existing version:

```
▶ benchcmp -best ~/sc1.bench ~/sc2.bench
benchmark                             old ns/op     new ns/op     delta
BenchmarkReplaceShortcodeTokens-4     20034         17970         -10.30%

benchmark                             old allocs     new allocs     delta
BenchmarkReplaceShortcodeTokens-4     1              11             +1000.00%

benchmark                             old bytes     new bytes     delta
BenchmarkReplaceShortcodeTokens-4     3072          12968         +322.14%
```