@bogem 2017-05-08
It's really strange. I become worse results with my branch on this benchmark too.
But the profiling of `hugo -s docs` give me opposite results, you can try it yourself:
Before:
3.58% of total
```
$ go tool pprof $GOPATH/bin/hugo hugo.mem
Entering interactive mode (type "help" for commands)
(pprof) web
(pprof) list replaceShortcode
Total: 88.65MB
ROUTINE ======================== github.com/spf13/hugo/hugolib.replaceShortcodeTokens in /Users/albert/go/src/github.com/spf13/hugo/hugolib/shortcode.go
    3.18MB     3.18MB (flat, cum)  3.58% of Total
         .          .    554:				end += 4
         .          .    555:			}
         .          .    556:		}
         .          .    557:
         .          .    558:		// This and other cool slice tricks: https://github.com/golang/go/wiki/SliceTricks
    3.18MB     3.18MB    559:		source = append(source[:j], append(newVal, source[end:]...)...)
         .          .    560:		start = j
         .          .    561:		k = bytes.Index(source[start:], pre)
         .          .    562:
         .          .    563:	}
         .          .    564:
```
After:
0.56% of total
```
$ go tool pprof -alloc_space $GOPATH/bin/hugo hugo.mem
Entering interactive mode (type "help" for commands)
(pprof) web
(pprof) list replaceShortcode
Total: 1.35GB
ROUTINE ======================== github.com/spf13/hugo/hugolib.replaceShortcodeTokens in /Users/albert/go/src/github.com/spf13/hugo/hugolib/shortcode.go
    4.08MB     7.70MB (flat, cum)  0.56% of Total
         .          .    542:			return nil, errors.New("illegal state in content: shortcode token missing end delim")
         .          .    543:		}
         .          .    544:
         .          .    545:		shortcodeEnd := preIndex + postIndex + len(post)
         .          .    546:
         .     1.50MB    547:		replacement := []byte(replacements[string(source[preIndex:shortcodeEnd])])
         .          .    548:		// If replacement also contains shortcodes.
         .          .    549:		if bytes.Contains(replacement, pre) {
         .          .    550:			var err error
         .          .    551:			replacement, err = replaceShortcodeTokens(replacement, replacements)
         .          .    552:			if err != nil {
         .          .    553:				return nil, err
         .          .    554:			}
         .          .    555:		}
         .          .    556:
         .          .    557:		// Issue #1148: Check for wrapping p-tags <p>
         .          .    558:		containPs := false
         .          .    559:		if preIndex >= len(pStart) && bytes.Equal(source[preIndex-len(pStart):preIndex], pStart) {
         .     1.07MB    560:			if preIndex+len(pEnd) < len(source) && bytes.Equal(source[shortcodeEnd:shortcodeEnd+len(pEnd)], pEnd) {
         .          .    561:				containPs = true
         .          .    562:			}
         .          .    563:		}
         .          .    564:
         .          .    565:		if containPs {
         .          .    566:			// Don't write <p>.
         .          .    567:			buf.Write(source[lastShortcodeEnd : preIndex-len(pStart)])
         .          .    568:			// Skip </p>.
         .          .    569:			shortcodeEnd += len(pEnd)
         .          .    570:		} else {
         .          .    571:			buf.Write(source[lastShortcodeEnd:preIndex])
         .          .    572:		}
         .          .    573:		buf.Write(replacement)
         .          .    574:
         .          .    575:		lastShortcodeEnd = shortcodeEnd
         .   540.51kB    576:		nextPre := bytes.Index(source[shortcodeEnd:], pre)
         .          .    577:		if nextPre < 0 {
         .          .    578:			break
         .          .    579:		} else {
         .          .    580:			preIndex = shortcodeEnd + nextPre
         .          .    581:		}
         .          .    582:	}
         .          .    583:
         .          .    584:	buf.Write(source[lastShortcodeEnd:])
    4.08MB     4.61MB    585:	source = make([]byte, buf.Len())
         .          .    586:	copy(source, buf.Bytes())
         .          .    587:
         .          .    588:	return source, nil
         .          .    589:}
         .          .    590:
```