@bogem 2017-05-08
You can benchmark by yourself with master and this branch with `hugo -s docs`.