@bep 2017-05-08
You are running pprof with two totally different settings, which is hardly fair game.

Note that by default:

```
- inuse_space -- a profile of the currently live objects/bytes on the heap 
```

Which is something completely different than:

```
go tool pprof -alloc_space
```