@bep 2017-05-08
>No, it should not be alloc free.

OK, so if not free -- then cheaper than a full copy.

My other benchmark (which includes the Hugo docs):

```
▶  benchcmp -best /tmp/1.bench /tmp/2.bench
benchmark           old ns/op       new ns/op       delta
BenchmarkHugo-4     11140334885     11374034739     +2.10%

benchmark           old allocs     new allocs     delta
BenchmarkHugo-4     53340138       53319790       -0.04%

benchmark           old bytes       new bytes       delta
BenchmarkHugo-4     11390503840     11285174072     -0.92%
```