@bogem 2017-05-31
Benchmarks are slightly the same with one shortcode:
```
name                                                                                                             old time/op    new time/op    delta
SiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=10,tags_per_page=5,shortcodes=true,render=true-4       6.86ms ± 2%    7.06ms ± 3%   ~     (p=0.400 n=3+3)
SiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=100,tags_per_page=5,shortcodes=true,render=true-4      47.6ms ± 1%    48.7ms ± 1%   ~     (p=0.100 n=3+3)
SiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=500,tags_per_page=5,shortcodes=true,render=true-4       208ms ± 2%     204ms ± 1%   ~     (p=0.400 n=3+3)
SiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=1000,tags_per_page=5,shortcodes=true,render=true-4      397ms ± 0%     396ms ± 3%   ~     (p=0.700 n=3+3)
SiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=5000,tags_per_page=5,shortcodes=true,render=true-4      1.84s ± 1%     1.89s ± 3%   ~     (p=0.200 n=3+3)
SiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=10000,tags_per_page=5,shortcodes=true,render=true-4     3.86s ± 3%     3.95s ± 7%   ~     (p=0.700 n=3+3)
SiteBuilding/frontmatter=TOML,num_root_sections=5,num_pages=10,tags_per_page=5,shortcodes=true,render=true-4       7.89ms ± 1%    7.69ms ± 1%   ~     (p=0.100 n=3+3)
SiteBuilding/frontmatter=TOML,num_root_sections=5,num_pages=100,tags_per_page=5,shortcodes=true,render=true-4      48.9ms ± 1%    48.9ms ± 1%   ~     (p=1.000 n=3+3)
SiteBuilding/frontmatter=TOML,num_root_sections=5,num_pages=500,tags_per_page=5,shortcodes=true,render=true-4       204ms ± 1%     206ms ± 2%   ~     (p=0.700 n=3+3)
SiteBuilding/frontmatter=TOML,num_root_sections=5,num_pages=1000,tags_per_page=5,shortcodes=true,render=true-4      388ms ± 1%     396ms ± 3%   ~     (p=0.400 n=3+3)
SiteBuilding/frontmatter=TOML,num_root_sections=5,num_pages=5000,tags_per_page=5,shortcodes=true,render=true-4      1.87s ± 3%     1.87s ± 2%   ~     (p=1.000 n=3+3)
SiteBuilding/frontmatter=TOML,num_root_sections=5,num_pages=10000,tags_per_page=5,shortcodes=true,render=true-4     3.78s ± 1%     3.82s ± 1%   ~     (p=0.400 n=3+3)

name                                                                                                             old alloc/op   new alloc/op   delta
SiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=10,tags_per_page=5,shortcodes=true,render=true-4       1.75MB ± 1%    1.76MB ± 1%   ~     (p=0.400 n=3+3)
SiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=100,tags_per_page=5,shortcodes=true,render=true-4      15.7MB ± 0%    15.8MB ± 1%   ~     (p=0.200 n=3+3)
SiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=500,tags_per_page=5,shortcodes=true,render=true-4      80.2MB ± 3%    79.9MB ± 1%   ~     (p=1.000 n=3+3)
SiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=1000,tags_per_page=5,shortcodes=true,render=true-4      162MB ± 0%     162MB ± 2%   ~     (p=1.000 n=3+3)
SiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=5000,tags_per_page=5,shortcodes=true,render=true-4      796MB ± 3%     807MB ± 1%   ~     (p=0.700 n=3+3)
SiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=10000,tags_per_page=5,shortcodes=true,render=true-4    1.68GB ± 0%    1.68GB ± 0%   ~     (p=0.400 n=3+3)
SiteBuilding/frontmatter=TOML,num_root_sections=5,num_pages=10,tags_per_page=5,shortcodes=true,render=true-4       1.91MB ± 0%    1.92MB ± 0%   ~     (p=0.100 n=3+3)
SiteBuilding/frontmatter=TOML,num_root_sections=5,num_pages=100,tags_per_page=5,shortcodes=true,render=true-4      15.9MB ± 1%    15.9MB ± 1%   ~     (p=1.000 n=3+3)
SiteBuilding/frontmatter=TOML,num_root_sections=5,num_pages=500,tags_per_page=5,shortcodes=true,render=true-4      79.2MB ± 3%    80.1MB ± 0%   ~     (p=1.000 n=3+3)
SiteBuilding/frontmatter=TOML,num_root_sections=5,num_pages=1000,tags_per_page=5,shortcodes=true,render=true-4      162MB ± 2%     158MB ± 2%   ~     (p=0.200 n=3+3)
SiteBuilding/frontmatter=TOML,num_root_sections=5,num_pages=5000,tags_per_page=5,shortcodes=true,render=true-4      799MB ± 2%     805MB ± 1%   ~     (p=0.700 n=3+3)
SiteBuilding/frontmatter=TOML,num_root_sections=5,num_pages=10000,tags_per_page=5,shortcodes=true,render=true-4    1.63GB ± 1%    1.63GB ± 2%   ~     (p=1.000 n=3+3)

name                                                                                                             old allocs/op  new allocs/op  delta
SiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=10,tags_per_page=5,shortcodes=true,render=true-4        24.9k ± 1%     24.9k ± 0%   ~     (p=0.700 n=3+3)
SiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=100,tags_per_page=5,shortcodes=true,render=true-4        179k ± 0%      180k ± 0%   ~     (p=0.400 n=3+3)
SiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=500,tags_per_page=5,shortcodes=true,render=true-4        863k ± 0%      864k ± 0%   ~     (p=1.000 n=3+3)
SiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=1000,tags_per_page=5,shortcodes=true,render=true-4      1.72M ± 0%     1.73M ± 1%   ~     (p=0.200 n=3+3)
SiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=5000,tags_per_page=5,shortcodes=true,render=true-4      8.58M ± 0%     8.59M ± 0%   ~     (p=0.100 n=3+3)
SiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=10000,tags_per_page=5,shortcodes=true,render=true-4     17.2M ± 0%     17.2M ± 0%   ~     (p=0.200 n=3+3)
SiteBuilding/frontmatter=TOML,num_root_sections=5,num_pages=10,tags_per_page=5,shortcodes=true,render=true-4        28.2k ± 0%     28.2k ± 0%   ~     (p=0.700 n=3+3)
SiteBuilding/frontmatter=TOML,num_root_sections=5,num_pages=100,tags_per_page=5,shortcodes=true,render=true-4        182k ± 0%      182k ± 0%   ~     (p=1.000 n=3+3)
SiteBuilding/frontmatter=TOML,num_root_sections=5,num_pages=500,tags_per_page=5,shortcodes=true,render=true-4        866k ± 0%      867k ± 0%   ~     (p=0.400 n=3+3)
SiteBuilding/frontmatter=TOML,num_root_sections=5,num_pages=1000,tags_per_page=5,shortcodes=true,render=true-4      1.72M ± 0%     1.72M ± 0%   ~     (p=0.700 n=3+3)
SiteBuilding/frontmatter=TOML,num_root_sections=5,num_pages=5000,tags_per_page=5,shortcodes=true,render=true-4      8.58M ± 0%     8.59M ± 0%   ~     (p=0.400 n=3+3)
SiteBuilding/frontmatter=TOML,num_root_sections=5,num_pages=10000,tags_per_page=5,shortcodes=true,render=true-4     17.2M ± 1%     17.2M ± 0%   ~     (p=1.000 n=3+3)
```
But with one more shortcode we can already see a benefit:
```
$ benchstat before.hugo after.hugo
name                                                                                                             old time/op    new time/op    delta
SiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=10,tags_per_page=5,shortcodes=true,render=true-4       7.96ms ± 6%    7.08ms ± 2%  -11.00%  (p=0.029 n=4+4)
SiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=100,tags_per_page=5,shortcodes=true,render=true-4      52.1ms ± 9%    49.5ms ± 2%     ~     (p=0.057 n=4+4)
SiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=500,tags_per_page=5,shortcodes=true,render=true-4       214ms ± 2%     214ms ± 3%     ~     (p=1.000 n=4+4)
SiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=1000,tags_per_page=5,shortcodes=true,render=true-4      412ms ± 0%     409ms ± 2%     ~     (p=0.343 n=4+4)
SiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=5000,tags_per_page=5,shortcodes=true,render=true-4      1.97s ± 2%     1.92s ± 2%     ~     (p=0.114 n=4+4)
SiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=10000,tags_per_page=5,shortcodes=true,render=true-4     4.12s ±11%     4.01s ± 3%     ~     (p=0.886 n=4+4)
SiteBuilding/frontmatter=TOML,num_root_sections=5,num_pages=10,tags_per_page=5,shortcodes=true,render=true-4       7.98ms ± 2%    7.81ms ± 1%     ~     (p=0.200 n=4+4)
SiteBuilding/frontmatter=TOML,num_root_sections=5,num_pages=100,tags_per_page=5,shortcodes=true,render=true-4      50.3ms ± 1%    50.8ms ± 2%     ~     (p=0.343 n=4+4)
SiteBuilding/frontmatter=TOML,num_root_sections=5,num_pages=500,tags_per_page=5,shortcodes=true,render=true-4       211ms ± 1%     213ms ± 1%     ~     (p=0.200 n=4+4)
SiteBuilding/frontmatter=TOML,num_root_sections=5,num_pages=1000,tags_per_page=5,shortcodes=true,render=true-4      410ms ± 1%     408ms ± 2%     ~     (p=0.686 n=4+4)
SiteBuilding/frontmatter=TOML,num_root_sections=5,num_pages=5000,tags_per_page=5,shortcodes=true,render=true-4      1.96s ± 2%     1.94s ± 2%     ~     (p=0.486 n=4+4)
SiteBuilding/frontmatter=TOML,num_root_sections=5,num_pages=10000,tags_per_page=5,shortcodes=true,render=true-4     3.92s ± 1%     3.95s ± 2%     ~     (p=0.486 n=4+4)

name                                                                                                             old alloc/op   new alloc/op   delta
SiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=10,tags_per_page=5,shortcodes=true,render=true-4       1.81MB ± 1%    1.80MB ± 1%     ~     (p=0.200 n=4+4)
SiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=100,tags_per_page=5,shortcodes=true,render=true-4      16.4MB ± 1%    16.2MB ± 1%     ~     (p=0.343 n=4+4)
SiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=500,tags_per_page=5,shortcodes=true,render=true-4      82.3MB ± 1%    82.2MB ± 0%     ~     (p=0.886 n=4+4)
SiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=1000,tags_per_page=5,shortcodes=true,render=true-4      167MB ± 3%     168MB ± 0%     ~     (p=0.343 n=4+4)
SiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=5000,tags_per_page=5,shortcodes=true,render=true-4      829MB ± 2%     826MB ± 3%     ~     (p=0.886 n=4+4)
SiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=10000,tags_per_page=5,shortcodes=true,render=true-4    1.72GB ± 1%    1.71GB ± 2%     ~     (p=0.343 n=4+4)
SiteBuilding/frontmatter=TOML,num_root_sections=5,num_pages=10,tags_per_page=5,shortcodes=true,render=true-4       1.96MB ± 1%    1.95MB ± 0%     ~     (p=0.200 n=4+4)
SiteBuilding/frontmatter=TOML,num_root_sections=5,num_pages=100,tags_per_page=5,shortcodes=true,render=true-4      16.4MB ± 0%    16.2MB ± 1%   -1.32%  (p=0.029 n=4+4)
SiteBuilding/frontmatter=TOML,num_root_sections=5,num_pages=500,tags_per_page=5,shortcodes=true,render=true-4      81.7MB ± 1%    82.0MB ± 1%     ~     (p=0.686 n=4+4)
SiteBuilding/frontmatter=TOML,num_root_sections=5,num_pages=1000,tags_per_page=5,shortcodes=true,render=true-4      166MB ± 1%     166MB ± 2%     ~     (p=1.000 n=4+4)
SiteBuilding/frontmatter=TOML,num_root_sections=5,num_pages=5000,tags_per_page=5,shortcodes=true,render=true-4      828MB ± 3%     823MB ± 3%     ~     (p=1.000 n=4+4)
SiteBuilding/frontmatter=TOML,num_root_sections=5,num_pages=10000,tags_per_page=5,shortcodes=true,render=true-4    1.70GB ± 2%    1.68GB ± 1%     ~     (p=0.343 n=4+4)

name                                                                                                             old allocs/op  new allocs/op  delta
SiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=10,tags_per_page=5,shortcodes=true,render=true-4        25.6k ± 1%     25.6k ± 1%     ~     (p=0.486 n=4+4)
SiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=100,tags_per_page=5,shortcodes=true,render=true-4        186k ± 1%      186k ± 1%     ~     (p=0.886 n=4+4)
SiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=500,tags_per_page=5,shortcodes=true,render=true-4        897k ± 1%      899k ± 1%     ~     (p=0.486 n=4+4)
SiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=1000,tags_per_page=5,shortcodes=true,render=true-4      1.79M ± 0%     1.79M ± 1%     ~     (p=0.200 n=4+4)
SiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=5000,tags_per_page=5,shortcodes=true,render=true-4      8.93M ± 0%     8.94M ± 0%     ~     (p=0.686 n=4+4)
SiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=10000,tags_per_page=5,shortcodes=true,render=true-4     17.9M ± 0%     17.9M ± 1%     ~     (p=0.343 n=4+4)
SiteBuilding/frontmatter=TOML,num_root_sections=5,num_pages=10,tags_per_page=5,shortcodes=true,render=true-4        28.8k ± 0%     28.8k ± 0%     ~     (p=0.686 n=4+4)
SiteBuilding/frontmatter=TOML,num_root_sections=5,num_pages=100,tags_per_page=5,shortcodes=true,render=true-4        188k ± 0%      188k ± 0%     ~     (p=1.000 n=4+4)
SiteBuilding/frontmatter=TOML,num_root_sections=5,num_pages=500,tags_per_page=5,shortcodes=true,render=true-4        900k ± 1%      902k ± 1%     ~     (p=0.686 n=4+4)
SiteBuilding/frontmatter=TOML,num_root_sections=5,num_pages=1000,tags_per_page=5,shortcodes=true,render=true-4      1.79M ± 0%     1.80M ± 0%   +0.50%  (p=0.029 n=4+4)
SiteBuilding/frontmatter=TOML,num_root_sections=5,num_pages=5000,tags_per_page=5,shortcodes=true,render=true-4      8.94M ± 0%     8.97M ± 0%     ~     (p=0.486 n=4+4)
SiteBuilding/frontmatter=TOML,num_root_sections=5,num_pages=10000,tags_per_page=5,shortcodes=true,render=true-4     17.9M ± 0%     17.9M ± 1%     ~     (p=0.686 n=4+4)
```