@bep 2017-06-01
You are persistent, I will say that. 

So, the benchmark that tests this specific func:

```
benchmark                             old ns/op     new ns/op     delta
BenchmarkReplaceShortcodeTokens-4     6284          5931          -5.62%

benchmark                             old allocs     new allocs     delta
BenchmarkReplaceShortcodeTokens-4     1              11             +1000.00%

benchmark                             old bytes     new bytes     delta
BenchmarkReplaceShortcodeTokens-4     3072          12978         +322.46%
```

Running `/benchSite.sh "YAML,num_pages=1000,shortcodes"` (which is 2 shortcodes per page on average, which is just about the same as the docs site):

```
benchmark                                                  old ns/op     new ns/op     delta
BenchmarkSiteBuilding/YAML,num_pages=1000,shortcodes-4     129728033     125539496     -3.23%

benchmark                                                  old allocs     new allocs     delta
BenchmarkSiteBuilding/YAML,num_pages=1000,shortcodes-4     674331         676327         +0.30%

benchmark                                                  old bytes     new bytes     delta
BenchmarkSiteBuilding/YAML,num_pages=1000,shortcodes-4     78728167      79307637      +0.74%
```

You can probably adjust the benchmark to make your func look better, but that isn't very interesting.

As to "easy to read", that is in the eye of the beholder. I see some tests that are changed, which does not look good, and something that I would have to look into if I'd consider merging this. All in all you have not convinced me to even start on that task when the main sell was performance.